requirements.txt
sentence-transformers==3.3.1
huggingface_hub
python-dotenv==1.0.0
pinecone-client==5.0.1
fastapi==0.95.1
uvicorn==0.20.0
python-multipart
tqdm
pandas
groq
